Averaging forecasts

To exploit the strong predictive performance of combining predictions from different forecasting models, PostForecasts.jl provides averaging schemes for both point and probabilistic forecasts:

PostForecasts.average — Function

average(pf; agg::Symbol=:mean)

Average the pool of point pred from pf. Return PointForecasts containing averaged forecasts, keyword argument agg specifies whether to average using simple mean (:mean) or median (:median).

If pf is a vector of PointForecasts, all individual forecasts from passed PointForecasts are averaged.

PostForecasts.qaverage — Function

qaverage(QF::Vector{QuantForecasts})

Average probabilistic pred from QF::Vector{QuantForecasts} by averaging quantiles of the distributions.

Return ProbcastsSeries containing quantile pred at the same prob as QuantForecasts stored in QF.

PostForecasts.paverage — Function

paverage(QF::Vector{QuantForecasts}, prob)

Average probabilistic pred from QF by averaging probabilities of the distributions.

Return ProbcastsSeries containing quantile pred at specified probabilities prob (vector of probabilities ::Vector{<:AbstractFloat}, single probability value ::AbstractFloat or the number of equidistant probability values ::Integer).