Forecasts structures

To make working with forecasts easy and user-friendy, PostForecasts.jl introduces Forecasts type, a supertype that spans PointForecasts and QuantForecast structures.

PointForecasts

PointForecasts is a structure intended for storing the series of point predictions (both single pred and pools of forecasts), along with the observations and identifiers (timestamps). The package provides functions for building PointForecasts objects from delimited files, calculating error measures (MAE and RMSE) and averaging point pred.

QuantForecasts

QuantForecasts struct is inteded for storing the series of probabilistic predictions, represented as quantiles of predictive distribution corresponding to probability levels, along with the observations and identifiers (timestamps). The package provides functions for computing probabilstic forecasts from PointForecasts objects, calculating pinball loss and averaging distributions across quantiles or probabilities.

Position-based indexing and slicing

PointForecasts and QuantForecasts support both position-based indexing and slicing. Accessing a series with a scalar index results in a named tuple, while slicing creates a new series object built from pred,observations and identifiers stored at respective indices.

pf = loaddata(:epex1);
firstday = pf[1]
#(pred = [27.640966097698737, 24.423563275081627, 23.54144377224293, 25.061033846927558], obs = 10.07, id = 20190101)
firstweek = pf[1:7]
#PointForecasts{Float64, Int64} with a pool of 4 forecasts at 7 timesteps, between 20190101 and 20190107

Label-based indexing

Since PointForecasts and QuantForecasts structures have id field storing an integer identifier for every timestep, it is posibble to access its elements by providng its identifier value, much in the spirit of Pandas .loc. Use () for label-based indexing. Analogously to indexing and slicing, providing a single label results in a named tuple, while a vector creates a new series object.

pf = loaddata(:epex1);
firstday = pf(20190101)
#(pred = [27.640966097698737, 24.423563275081627, 23.54144377224293, 25.061033846927558], obs = 10.07, id = 20190101)
firstweek = pf([20190101, 20190102, 20190103, 20190104, 20190105, 20190106, 20190107])
#PointForecasts{Float64, Int64} with a pool of 4 forecasts at 7 timesteps, between 20190101 and 20190107

PostForecasts.PointForecasts — Type

PointForecasts(pred, obs, id)

Create PointForecasts<:Forecasts for storing the series of point predictions, along with the observations and identifiers.

Be aware that the constructor does not copy the arguments.

PostForecasts.QuantForecasts — Type

QuantForecasts(pred, obs, id, prob)

Create QuantForecasts<:Forecasts for storing the series of probabilistic predictions, represented as quantiles of predictive distribution at specfied probabilities, along with the observations and identifiers.

Be aware that the constructor does not copy the arguments.

PostForecasts.findindex — Function

findindex(fs::Forecasts, i::Integer)

Return the index of fs, for which the element of field id equals i.

PostForecasts.decouple — Function

decouple(pf::PointForecasts)

Return ::Vector{PointForecasts}, where each element contains an individual forecast series from pf.

PostForecasts.npred — Function

npred(fs::Forecasts)

Return the number of point forecasts in fs::PointForecasts or the number of forecasted quantiles in fs::QuantForecasts.

PostForecasts.setpred — Function

setpred(fs::Forecasts, t::Integer, i::Integer, val::AbstractFloat)

Set the element of field fs.pred at indices t, i to val.

PostForecasts.getpred — Function

getpred(fs::Forecasts[, T, I])

Return the copy of predictions from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get predictions at specified time indices.

Additionally, provide I::Union{Integer, AbstractVector{<:Integer}} to get predicitons at specified forecast indices.

PostForecasts.getobs — Function

getobs(fs::Forecasts[, T])

Return the copy of observations from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get observations at specified time indices.

PostForecasts.getid — Function

getid(fs::Forecasts[, T])

Return the copy of identifiers from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get identifiers at specified time indices.

PostForecasts.getprob — Function

getprob(qf::QuantForecasts[, I])

Return the copy of probabilities from qf.

Provide optional argument I::Union{Integer, AbstractVector{<:Integer}} to get probabilities at specified forecast indices.

PostForecasts.viewpred — Function

viewpred(fs::Forecasts[, T, I])

Return the view of predictions from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get predictions at specified time indices.

Additionally, provide I::Union{Integer, AbstractVector{<:Integer}} to get predicitons at specified forecast indices.

PostForecasts.viewobs — Function

viewobs(fs::Forecasts[, T])

Return the view of observations from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get observations at specified time indices.

PostForecasts.viewid — Function

viewid(fs::Forecasts[, T])

Return the view of identifiers from fs.

Provide optional argument T::Union{Integer, AbstractVector{<:Integer}} to get identifiers at specified time indices.

PostForecasts.viewprob — Function

viewprob(qf::QuantForecasts[, I])

Return the view of probabilities from qf.

Provide optional argument I::Union{Integer, AbstractVector{<:Integer}} to get probabilities at specified forecast indices.