Loading and saving forecasts

With PostForecasts.jl You can easily create PointForecasts from delimited files, load and save both PointForecasts and QuantForecast using HDF5 format and play with pre-installed datasets.

To make managing files generated with PostForecasts.jl easier, HDF5 files containing PointForecasts and QuantForecasts are saved with .pointf and .quantf extensions respectively.

PostForecasts.loaddata — Function

loaddata(dataset::Symbol)

Create a PointForecasts object from the dataset provided with the package, availabe options include:

:epexH, where H is an integer between 1 and 24

:pangu'H'u10, where H is an integer between 0 and 186, divisible by 6.

:pangu'H'v10, where H is an integer between 0 and 186, divisible by 6.

:pangu'H't2m, where H is an integer between 0 and 186, divisible by 6.

:pangu'H't850, where H is an integer between 0 and 186, divisible by 6.

:pangu'H'z500, where H is an integer between 0 and 186, divisible by 6.

Details of the datasets are avaiable in documentation.

PostForecasts.loaddlmdata — Function

loaddlmdata(filepath::String; kwargs...)

Create a PointForecasts object from delimited file at filepath.

Keyword Arguments

delim=',': Specifies the delimitter

idcol=1: Specifies which column is used for timestamps (0 to generate timestamps automatically)

obscol=2: Specifies which column is used for observations

predcol=0: Specifies which columns are used for pred (0 to use all remaining columns)

colnames=false If true, omit the first row of the file.

PostForecasts.save — Function

save(f::Forecasts, filepath::String, groupname::String="forecasts")

Save f to a HDF5 file at filepath (.pointf or .quantf extension is added if missing). Optionally specify the name of the group, in which the data will be saved.

PostForecasts.loadpointf — Function

loadpointf(filepath::String, groupname::String="forecasts")

Load PointForecasts from a HDF5 file at filepath (.pointf extension is added if missing). Optionally specify the name of the group, from which the data will be loaded.

PostForecasts.loadquantf — Function

loadquantf(filepath::String, groupname::String="forecasts")

Load QuantForecasts from a HDF5 file at filepath (.quantf extension is added if missing). Optionally specify the name of the group, from which the data will be loaded.