Pre-release functionalities

Regularized quantile regressions

PostForecasts.iQR — Function

iQR(args...)
Create an isotonic quantile regression model, constraining the weights to be non-negative. The arguments `args...` are the same as for `QR`.

PostForecasts.LassoQR — Type

LassoQR([type::Type{F}=Float64,] n::Integer, r::Integer, prob::Union{AbstractFloat, AbstractVector{<:AbstractFloat}}, lambda::Union{AbstractFloat, AbstractVector{<:AbstractFloat}}) where {F<:AbstractFloat}

Creates a LassoQR{F}<:MultiPostModel{F}<:PostModel{F} model for lasso quantile regression with regularization strength lambda to be trained on n observations with r forecasts (regressors), fitting quantiles at probabilities specified by prob.

If lambda is a vector, the optimal regularization strength will be selected using the Bayesian Information Criterion (BIC) during every training, separately for each quantile.

By default, lambda is specified by the package constant LAMBDA = [0.001, 0.01, 0.1, 1.0, 10.0]. It can be modified with the setlambda method.

PostForecasts.setlambda — Function

Set the values of the package constant LAMBDA to be equal to lambda.

PostForecasts.getlambda — Function

Get the values stored in the package constant `LAMBDA`.