Utilities

A place to store useful functions that didn't fit elsewhere

PostForecasts.getmodel — Function

getmodel(::Val, params...)

Helper function that dispatches the model based on the model name passed as Val.

Available methods

getmodel(Val(:qra, n, m, prob), for Quantile Regression Averaging

getmodel(Val(:cp), n) for Conformal Prediction

getmodel(Val(:hs), n) for Conformal Prediction Prediction with non-symmetric errors (a.k.a. Historical Simulation)

getmodel(Val(:idr), n, m) for Isotonic Distributional Regression

getmodel(Val(:normal)) for Normal distribution of errors

getmodel(Val(:zeronormal)) for Normal distribution of errors with fixed mean equal to 0

where n is the length of the training window, m is the number of regressors and prob is the probability (scalar value or vector).

Return an appropriate ProbModel.

PostForecasts.nreg — Function

nreg(m::ProbModel)

Return the number of regressors of model m.

PostForecasts.matchwindow — Function

matchwindow(m::ProbModel, window::Integer)

Return true if window matches the specification of model m, otherwise return false.

PostForecasts.checkmatch — Function

chechmatch(S::AbstractVector{<:ForecastSeries}; checkpred::Bool=false)

Check if PointForecasts or QuantForecasts provided in vector S match, i.e.:

all their identifiers match

all their observations match

their number of forecasts match (if checkpred=true)

their probabilities match (for QuantForecasts if checkpred=true).

Return nothing, throw an ArgumentError if any of the requirements above is not met.