Forecast Averaging

Averaging schemes

To exploit the strong predictive performance of combining predictions from different forecasting models, PostForecasts.jl provides averaging schemes for both point and probabilistic forecasts:

PostForecasts.average — Function

average(pf; agg::Symbol=:mean)

Average the pool of point pred from pf. Return PointForecasts containing averaged forecasts, keyword argument agg specifies whether to average using simple mean (:mean) or median (:median).

Argument types

pf::PointForecasts to average the pool of forecats in pf

pf::PointForecasts... to average all individual forecasts from every PointForecasts in pf.

pf::AbstractVector{<:PointForecasts} to average all individual forecasts from every PointForecasts in pf.

PostForecasts.qaverage — Function

qaverage(qfs)

Average probabilistic predictions from a collection qfs by averaging the quantiles.

Return QuantForecasts containing quantile predictions at the same quantile levels as QuantForecasts in qfs.

PostForecasts.paverage — Function

paverage(qfs[; quantiles])

Average probabilistic predictions from a collection qfs by averaging the distributions across probability.

Return QuantForecasts containing predictions of specified quantiles:

quantiles::AbstractVector{<:AbstractFloat}: vector of probabilities

quantiles::AbstractFloat: a single probability value

quantiles::Integer: number of equidistant probability values (e.g. 99 for percentiles).