Loading and saving forecasts

With PostForecasts.jl You can easily create PointForecasts from delimited files, load and save both PointForecasts and QuantForecast using HDF5 format and play with pre-installed datasets.

To make managing files generated with PostForecasts.jl easier, HDF5 files containing PointForecasts and QuantForecasts are saved with .pointf and .quantf extensions respectively.

PostForecasts.loaddata — Function

loaddata(dataset::Union{Symbol, AbstractString})

Create a PointForecasts object from the dataset provided with the package, availabe options include:

epexH, where H is an integer between 1 and 24

pangu'H'u10, where H is an integer between 0 and 186, divisible by 6.

pangu'H'v10, where H is an integer between 0 and 186, divisible by 6.

pangu'H't2m, where H is an integer between 0 and 186, divisible by 6.

pangu'H't850, where H is an integer between 0 and 186, divisible by 6.

pangu'H'z500, where H is an integer between 0 and 186, divisible by 6.

Details of the datasets are available in documentation.

PostForecasts.loaddlm — Function

loaddlm(filepath::AbstractString; kwargs...)

Create a PointForecasts object from delimited file at filepath.

Keyword Arguments

delim=',': Specifies the delimitter

obscol=1: Specifies which column is used for observations

predcol=nothing: Specifies which columns are used for pred (omit to use all remaining columns)

idcol=nothing: Specifies which column is used for timestamps (omit to generate timestamps automatically)

colnames=false If true, omit the first row of the file.

PostForecasts.saveforecasts — Function

saveforecasts(f::Forecasts, filepath::AbstractString)

Save f to a HDF5 file at filepath with .pointf extension for PointForecasts and .quantf extension for QuantForecasts (extension is added if missing).

PostForecasts.loadforecasts — Function

loadforecasts(filepath::AbstractString)::Forecasts

Load PointForecasts or QuantForecasts from filepath (.pointf or .quantf extension is required).

PostForecasts.loadpointf — Function

loadpointf(filepath::AbstractString)::PointForecasts

Load PointForecasts from filepath (.pointf extension is added if missing).

PostForecasts.loadquantf — Function

loadquantf(filepath::AbstractString)::QuantForecasts

Load QuantForecasts from filepath (.quantf extension is added if missing).